feat: fetching script
