+++ server/config/db/versions/af37f8067e83_section_table_init.py
@@ -55,4 +55,4 @@ def upgrade() -> None:
 
 
 def downgrade() -> None:
-    pass
+    op.drop_table("sections")
